+++ DCONFIG_Utils.py
@@ -36,6 +36,11 @@ def active_mesh_available(context):
     return active_object is not None and active_object.type == 'MESH'
 
 
+def active_object_available(context, obj_types):
+    active_object = context.active_object
+    return active_object is not None and active_object.type in obj_types
+
+
 def active_mesh_selected(context):
     active_object = context.active_object
     return active_object is not None and active_object.type == 'MESH' and (context.mode == 'EDIT_MESH' or active_object.select_get())
